# src/web/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "BeeEye Platform API",
    "description": "AI decision-intelligence platform for ADMC. Read-only analytics in the initial implementation. All bounded contexts are mounted under /api/v1.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5080/"
    }
  ],
  "paths": {
...
    "/api/v1/predictions/explain": {
      "get": {
        "tags": [
          "Predictions"
        ],
        "summary": "Why the platform says what it says about one subject",
        "operationId": "Predictions_Explain",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplanationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/explain/feedback": {
      "post": {
        "tags": [
          "Predictions"
        ],
        "summary": "Record whether an explanation was useful. Recorded only; it retrains nothing",
        "operationId": "Predictions_ExplainFeedback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfidenceDto": {
        "required": [
          "band",
          "percent",
          "why"
        ],
        "type": "object",
        "properties": {
          "band": {
            "type": "string"
          },
          "percent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "why": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "DriverDto": {
        "required": [
          "label",
          "detail"
        ],
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "EvidencePointDto": {
        "required": [
          "label",
          "value",
          "comparison"
        ],
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "value": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "comparison": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "EvidenceSeriesDto": {
        "required": [
          "period",
          "points",
          "note",
          "valueLabel",
          "comparisonLabel"
        ],
        "type": "object",
        "properties": {
          "period": {
            "type": "string"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidencePointDto"
            }
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          },
          "valueLabel": {
            "type": "string"
          },
          "comparisonLabel": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ExplanationDto": {
        "required": [
          "title",
          "module",
          "label",
          "recommendation",
          "impacts",
          "confidence",
          "drivers",
          "evidence",
          "assumptions",
          "lineage",
          "model",
          "ownership",
          "isDemoData"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "module": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "recommendation": {
            "type": [
              "null",
              "string"
            ]
          },
          "impacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImpactTileDto"
            }
          },
          "confidence": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConfidenceDto"
              }
            ]
          },
          "drivers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DriverDto"
            }
          },
          "evidence": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EvidenceSeriesDto"
              }
            ]
          },
          "assumptions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lineage": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineageNodeDto"
            }
          },
          "model": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ModelInfoDto"
              }
            ]
          },
          "ownership": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OwnershipDto"
              }
            ]
          },
          "isDemoData": {
            "type": "boolean"
          }
        }
      },
      "ExplanationGapDto": {
        "required": [
          "area",
          "reason"
        ],
        "type": "object",
        "properties": {
          "area": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "ExplanationResponse": {
        "required": [
          "subjectKind",
          "subjectRef",
          "explanation",
          "gaps",
          "feedback",
          "feedbackCaveat",
          "generatedAtUtc"
        ],
        "type": "object",
        "properties": {
          "subjectKind": {
            "type": "string"
          },
          "subjectRef": {
            "type": "string"
          },
          "explanation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExplanationDto"
              }
            ]
          },
          "gaps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExplanationGapDto"
            }
          },
          "feedback": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedbackEntryDto"
            }
          },
          "feedbackCaveat": {
            "type": "string"
          },
          "generatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FeedbackEntryDto": {
        "required": [
          "verdict",
          "note",
          "submittedBy",
          "submittedAtUtc"
        ],
        "type": "object",
        "properties": {
          "verdict": {
            "type": "string"
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          },
          "submittedBy": {
            "type": "string"
          },
          "submittedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FeedbackRequest": {
        "required": [
          "kind",
          "ref",
          "verdict",
          "note"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "type": [
              "null",
              "string"
            ]
          },
          "ref": {
            "type": [
              "null",
              "string"
            ]
          },
          "verdict": {
            "type": [
              "null",
              "string"
            ]
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "FeedbackResponse": {
        "required": [
          "subjectKind",
          "subjectRef",
          "verdict",
          "submittedAtUtc",
          "caveat"
        ],
        "type": "object",
        "properties": {
          "subjectKind": {
            "type": "string"
          },
          "subjectRef": {
            "type": "string"
          },
          "verdict": {
            "type": "string"
          },
          "submittedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "caveat": {
            "type": "string"
          }
        }
      },
...
      "ImpactTileDto": {
        "required": [
          "label",
          "value",
          "tone"
        ],
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "tone": {
            "type": "string"
          }
        }
      },
...
      "LineageNodeDto": {
        "required": [
          "label",
          "kind"
        ],
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          }
        }
      },
...
      "ModelInfoDto": {
        "required": [
          "name",
          "version",
          "recalculated",
          "horizon",
          "validation",
          "error"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "recalculated": {
            "type": "string"
          },
          "horizon": {
            "type": "string"
          },
          "validation": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "OwnershipDto": {
        "required": [
          "ownerRole",
          "status",
          "decisionSubjectRef"
        ],
        "type": "object",
        "properties": {
          "ownerRole": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "decisionSubjectRef": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}